requirements.txt
folium
streamlit-folium
plotly
wikipedia